extractTimeFID function returns array of integer

+++ chemclipse/plugins/org.eclipse.chemclipse.nmr.model/src/org/eclipse/chemclipse/nmr/model/support/ISignalExtractor.java
@@ -37,7 +37,7 @@ public interface ISignalExtractor {
 
 	double[] extractIntesityFIDReal();
 
-	double[] extractTimeFID();
+	int[] extractTimeFID();
 
 	void createScansFID(Complex[] complexFID, int[] time);
